Add new module dbk_kernels_merging.py
