+++ diffsky/experimental/kernels/mc_phot_kernels.py
@@ -1,5 +1,4 @@
 from collections import namedtuple
-from functools import partial
 
 from dsps.sfh.diffburst import DEFAULT_BURST_PARAMS
 from jax import jit as jjit
@@ -9,7 +8,6 @@ from jax import vmap
 
 from ...merging import compute_x_tot_from_x_in_situ, merging_model
 from ...ssp_err_model import ssp_err_model
-from .. import mc_diffstarpop_wrappers as mcdw
 from ..disk_bulge_modeling import disk_bulge_kernels as dbk
 from . import (
     constants,
@@ -17,6 +15,7 @@ from . import (
     linelum_kernels,
     mc_randoms,
     phot_kernels,
+    phot_kernels_merging,
     sed_kernels,
 )
 from . import ssp_weight_kernels as sspwk
@@ -38,6 +37,9 @@ _specphot_kern = linelum_kernels._specphot_kern
 _get_dbk_phot_from_dbk_weights = dbk_kernels._get_dbk_phot_from_dbk_weights
 _sed_kern = sed_kernels._sed_kern
 _get_dbk_linelum_decomposition = dbk_kernels._get_dbk_linelum_decomposition
+_mc_phot_kern_merging = phot_kernels_merging._mc_phot_kern_merging
+_phot_kern_merging = phot_kernels_merging._phot_kern_merging
+
 
 # randoms
 get_mc_phot_randoms = mc_randoms.get_mc_phot_randoms
@@ -49,129 +51,6 @@ SpecKernResults = linelum_kernels.SpecKernResults
 PhotKernResults = phot_kernels.PhotKernResults
 DBKRandoms = mc_randoms.DBKRandoms
 SEDKernResults = sed_kernels.SEDKernResults
-
-
-@partial(jjit, static_argnames=["n_t_table"])
-def _mc_phot_kern_merging(
-    ran_key,
-    z_obs,
-    t_obs,
-    mah_params,
-    ssp_data,
-    precomputed_ssp_mag_table,
-    z_phot_table,
-    wave_eff_table,
-    diffstarpop_params,
-    mzr_params,
-    spspop_params,
-    scatter_params,
-    ssp_err_pop_params,
-    merge_params,
-    cosmo_params,
-    fb,
-    logmp_infall,
-    logmhost_infall,
-    t_infall,
-    is_central,
-    nhalos_weights,
-    halo_indx,
-    n_t_table=mcdw.N_T_TABLE,
-):
-    phot_randoms, sfh_params = get_mc_phot_randoms(
-        ran_key, diffstarpop_params, mah_params, cosmo_params
-    )
-    phot_kern_results, flux_obs, merge_prob, mstar_obs = _phot_kern_merging(
-        phot_randoms,
-        sfh_params,
-        z_obs,
-        t_obs,
-        mah_params,
-        ssp_data,
-        precomputed_ssp_mag_table,
-        z_phot_table,
-        wave_eff_table,
-        mzr_params,
-        spspop_params,
-        scatter_params,
-        ssp_err_pop_params,
-        merge_params,
-        cosmo_params,
-        fb,
-        logmp_infall,
-        logmhost_infall,
-        t_infall,
-        is_central,
-        nhalos_weights,
-        halo_indx,
-        n_t_table=n_t_table,
-    )
-    return phot_kern_results, phot_randoms, flux_obs, merge_prob, mstar_obs
-
-
-@partial(jjit, static_argnames=["n_t_table"])
-def _phot_kern_merging(
-    phot_randoms,
-    sfh_params,
-    z_obs,
-    t_obs,
-    mah_params,
-    ssp_data,
-    precomputed_ssp_mag_table,
-    z_phot_table,
-    wave_eff_table,
-    mzr_params,
-    spspop_params,
-    scatter_params,
-    ssp_err_pop_params,
-    merge_params,
-    cosmo_params,
-    fb,
-    logmp_infall,
-    logmhost_infall,
-    t_infall,
-    is_central,
-    nhalos_weights,
-    halo_indx,
-    n_t_table=mcdw.N_T_TABLE,
-):
-    phot_kern_results = _phot_kern(
-        phot_randoms,
-        sfh_params,
-        z_obs,
-        t_obs,
-        mah_params,
-        ssp_data,
-        precomputed_ssp_mag_table,
-        z_phot_table,
-        wave_eff_table,
-        mzr_params,
-        spspop_params,
-        scatter_params,
-        ssp_err_pop_params,
-        cosmo_params,
-        fb,
-        n_t_table=n_t_table,
-    )
-
-    upids = jnp.where(is_central == 1, -1.0, 0.0)
-    merge_prob = merging_model.get_p_merge_from_merging_params(
-        merge_params, logmp_infall, logmhost_infall, t_obs, t_infall, upids
-    )
-
-    mstar_in_situ = 10**phot_kern_results.logsm_obs
-    mstar_obs = compute_x_tot_from_x_in_situ(
-        mstar_in_situ, merge_prob, nhalos_weights, halo_indx
-    )
-
-    flux_in_situ = 10 ** (-0.4 * phot_kern_results.obs_mags)
-    flux_obs = compute_x_tot_from_x_in_situ(
-        flux_in_situ,
-        merge_prob[:, jnp.newaxis],
-        nhalos_weights[:, jnp.newaxis],
-        halo_indx,
-    )
-
-    return phot_kern_results, flux_obs, merge_prob, mstar_obs
 
 
 @jjit
